Prevent empty content from causing build failures.
Using a new Gatsby core API, we can explicitly define the data types
for Markdown frontmatter fields we expect to exist (rather than rely
on Gatby's GraphQL schema inferencing). Previously, no content meant
there was nothing for Gatsby to make an inference based on, and it
simply failed to build.